fix text field color in all search screens and login screen and make card click

+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/searchcountry/SearchByCountryInteractionListener.kt
@@ -5,5 +5,5 @@ interface SearchByCountryInteractionListener {
     fun onCountryClicked()
     fun onDismissDropDown()
     fun onBackClicked()
-    fun onMovieClicked(movieId: Int)
+    fun onMovieClicked(movieId: Long)
 }
